# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Indian Union Budget Analysis Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface YearSummary {
  year: string;
  total_schemes: number;
  total_rows: number;
  total_budget_value: number;
  top_sector: string;
  anomaly_count: number;
  accuracy_score: number;
}

export interface SectorAllocation {
  sector: string;
  total_value: number;
  row_count: number;
}

export interface TrendDataPoint {
  year: string;
  total_value: number;
  row_count: number;
}

export type HeatmapDataMatrix = {[key: string]: {[key: string]: number}};

export interface HeatmapData {
  sectors: string[];
  years: string[];
  matrix: HeatmapDataMatrix;
}

export interface AnomalyRecord {
  scheme_name: string;
  sector: string;
  source: string;
  /** @nullable */
  value: number | null;
  /** @nullable */
  context?: string | null;
}

export interface SchemeRecord {
  scheme_name: string;
  sector: string;
  /** @nullable */
  source?: string | null;
  /** @nullable */
  value: number | null;
}

export interface SchemeList {
  items: SchemeRecord[];
  total: number;
  page: number;
  page_size: number;
}

export interface AccuracyMetrics {
  accuracy: number;
  valid_names_pct: number;
  valid_numbers_pct: number;
  valid_sector_pct: number;
  clean_rows_pct: number;
}

export interface YearComparison {
  sector: string;
  year1_value: number;
  year2_value: number;
  change_pct: number;
}

export interface PipelineStatus {
  row_count: number;
  status: string;
}

export type GetSectorAllocationParams = {
doc_type?: string;
};

export type GetSectorTrendParams = {
doc_type?: string;
};

export type GetHeatmapParams = {
doc_type?: string;
};

export type GetAnomaliesParams = {
/**
 * @nullable
 */
year?: string | null;
/**
 * @nullable
 */
sector?: string | null;
limit?: number;
};

export type GetSchemesParams = {
/**
 * @nullable
 */
year?: string | null;
/**
 * @nullable
 */
sector?: string | null;
/**
 * @nullable
 */
search?: string | null;
page?: number;
page_size?: number;
};

export type GetAccuracyParams = {
/**
 * @nullable
 */
year?: string | null;
};

export type CompareTwoYearsParams = {
year1: string;
year2: string;
doc_type?: string;
};

export type GetTopSchemesParams = {
top_n?: number;
/**
 * @nullable
 */
sector?: string | null;
};